// src/generated/SwapPool.client.ts
export interface SwapPoolReadOnlyInterface {
  contractAddress: string;
  balance: ({
    address
  }: {
    address: string;
  }) => Promise<BalanceResponse>;
  info: () => Promise<InfoResponse>;
  token1ForToken2Price: ({
    token1Amount
  }: {
    token1Amount: Uint128;
  }) => Promise<Token1ForToken2PriceResponse>;
  token2ForToken1Price: ({
    token2Amount
  }: {
    token2Amount: Uint128;
  }) => Promise<Token2ForToken1PriceResponse>;
}
export class SwapPoolQueryClient implements SwapPoolReadOnlyInterface {
  client: CosmWasmClient;
  contractAddress: string;

  constructor(client: CosmWasmClient, contractAddress: string) {
    this.client = client;
    this.contractAddress = contractAddress;
    this.balance = this.balance.bind(this);
    this.info = this.info.bind(this);
    this.token1ForToken2Price = this.token1ForToken2Price.bind(this);
    this.token2ForToken1Price = this.token2ForToken1Price.bind(this);
  }

  balance = async ({
    address
  }: {
    address: string;
  }): Promise<BalanceResponse> => {
    return this.client.queryContractSmart(this.contractAddress, {
      balance: {
        address
      }
    });
  };
  info = async (): Promise<InfoResponse> => {
    return this.client.queryContractSmart(this.contractAddress, {
      info: {}
    });
  };
  token1ForToken2Price = async ({
    token1Amount
  }: {
    token1Amount: Uint128;
  }): Promise<Token1ForToken2PriceResponse> => {
    return this.client.queryContractSmart(this.contractAddress, {
      token1_for_token2_price: {
        token1_amount: token1Amount
      }
    });
  };
  token2ForToken1Price = async ({
    token2Amount
  }: {
    token2Amount: Uint128;
  }): Promise<Token2ForToken1PriceResponse> => {
    return this.client.queryContractSmart(this.contractAddress, {
      token2_for_token1_price: {
        token2_amount: token2Amount
      }
    });
  };
}